WW-5219 Moves TestNG related classes into TestNG plugin Also moves test classes under testng package
