+++ plugins/testng/pom.xml
@@ -35,7 +35,6 @@
         <dependency>
             <groupId>org.testng</groupId>
             <artifactId>testng</artifactId>
-            <version>6.9.10</version>
         </dependency>
         <dependency>
             <groupId>org.springframework</groupId>
